+++ src/main/java/net/grid/vampiresdelight/common/food/SpecialFoodProperties.java
@@ -1,4 +1,0 @@
-package net.grid.vampiresdelight.common.food;
-
-public class SpecialFoodProperties {
-}